change id in user to type Long

+++ src/main/java/com/example/controllers/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     }
 
     @GetMapping("/{userDtoId}")
-    public UserDto findById(@PathVariable int userDtoId) {
+    public UserDto findById(@PathVariable Long userDtoId) {
         return userService.findById(userDtoId);
     }
